pyproject.toml
[project]
name = "agent-container-pack"
version = "0.1.0"
description = "CLI tool for generating CLAUDE.md/AGENTS.md from agentpack.yml"
readme = "README.md"
authors = [{ name = "ryoooo" }]
license = { text = "MIT" }
keywords = ["claude-code", "codex", "devcontainer", "mcp", "ai-assistant"]
requires-python = ">=3.13"

dependencies = [
    "cyclopts>=4.4.3",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://github.com/ryoooo/agent-container-pack"
Repository = "https://github.com/ryoooo/agent-container-pack"

[project.scripts]
acp = "agent_container_pack.cli:app"

[dependency-groups]
dev = [
    "pre-commit",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff",
    "syrupy>=5.0.0",
    "ty",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_container_pack"]
